feat: add debug messaging for startup times
